Separate `Store` into smaller chunks
