+++ packages/graph/hash_graph/lib/graph/src/store/postgres/onthology/database_type.rs
@@ -3,30 +3,30 @@ use crate::ontology::types::{DataType, EntityType, LinkType, OntologyType, Prope
 /// Provides an abstraction over elements of the Type System stored in the Database.
 ///
 /// [`PostgresDatabase`]: crate::store::PostgresDatabase
-pub trait DatabaseType: OntologyType {
+pub trait OntologyDatabaseType: OntologyType {
     /// Returns the name of the table where this type is stored.
     fn table() -> &'static str;
 }
 
-impl DatabaseType for DataType {
+impl OntologyDatabaseType for DataType {
     fn table() -> &'static str {
         "data_types"
     }
 }
 
-impl DatabaseType for PropertyType {
+impl OntologyDatabaseType for PropertyType {
     fn table() -> &'static str {
         "property_types"
     }
 }
 
-impl DatabaseType for EntityType {
+impl OntologyDatabaseType for EntityType {
     fn table() -> &'static str {
         "entity_types"
     }
 }
 
-impl DatabaseType for LinkType {
+impl OntologyDatabaseType for LinkType {
     fn table() -> &'static str {
         "link_types"
     }
