+++ packages/graph/hash_graph/lib/graph/src/store/postgres/onthology/mod.rs
@@ -1,0 +1,4 @@
+mod database_type;
+mod read;
+
+pub use self::database_type::OntologyDatabaseType;
