+++ packages/graph/hash_graph/lib/graph/src/store/postgres/property_type/mod.rs
@@ -1,0 +1,76 @@
+use async_trait::async_trait;
+use error_stack::{IntoReport, Result, ResultExt};
+use tokio_postgres::GenericClient;
+
+use crate::{
+    ontology::AccountId,
+    store::{AsClient, InsertionError, PostgresStore, PropertyTypeStore, UpdateError},
+    PropertyType,
+};
+
+#[async_trait]
+impl<C: AsClient> PropertyTypeStore for PostgresStore<C> {
+    async fn create_property_type(
+        &mut self,
+        property_type: &PropertyType,
+        created_by: AccountId,
+    ) -> Result<(), InsertionError> {
+        let transaction = PostgresStore::new(
+            self.as_mut_client()
+                .transaction()
+                .await
+                .report()
+                .change_context(InsertionError)?,
+        );
+
+        let version_id = transaction.create(property_type, created_by).await?;
+
+        transaction
+            .insert_property_type_references(property_type, version_id)
+            .await
+            .change_context(InsertionError)
+            .attach_printable("Could not insert references for property type")
+            .attach_lazy(|| property_type.clone())?;
+
+        transaction
+            .client
+            .commit()
+            .await
+            .report()
+            .change_context(InsertionError)?;
+
+        Ok(())
+    }
+
+    async fn update_property_type(
+        &mut self,
+        property_type: &PropertyType,
+        updated_by: AccountId,
+    ) -> Result<(), UpdateError> {
+        let transaction = PostgresStore::new(
+            self.as_mut_client()
+                .transaction()
+                .await
+                .report()
+                .change_context(UpdateError)?,
+        );
+
+        let version_id = transaction.update(property_type, updated_by).await?;
+
+        transaction
+            .insert_property_type_references(property_type, version_id)
+            .await
+            .change_context(UpdateError)
+            .attach_printable("Could not insert references for property type")
+            .attach_lazy(|| property_type.clone())?;
+
+        transaction
+            .client
+            .commit()
+            .await
+            .report()
+            .change_context(UpdateError)?;
+
+        Ok(())
+    }
+}
